requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Vector Store & Embeddings
chromadb==0.4.22
openai==1.10.0
anthropic==0.18.1
tiktoken==0.5.2

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
python-multipart==0.0.6
markdown==3.5.2

# Audio Processing
openai-whisper==20231117

# Text Processing
langchain==0.1.6
langchain-openai==0.0.5
langchain-community==0.0.17

# NLP & Analysis
nltk==3.8.1
textstat==0.7.3
scikit-learn==1.4.0

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
tenacity==8.2.3

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

